requirements.txt
torch==1.13.1
Pillow
matplotlib
h5py
numpy
pandas
tqdm